:sparkles: feat: AgentViewer 抽象
