+++ src/features/AgentViewer/style.ts
@@ -1,7 +1,5 @@
 import { createStyles } from 'antd-style';
 
-import { HEADER_HEIGHT } from '@/constants/common';
-
 export const useStyles = createStyles(({ css, token }) => ({
   toolbar: css`
     position: absolute;
@@ -16,10 +14,8 @@ export const useStyles = createStyles(({ css, token }) => ({
   `,
   canvas: css`
     display: block;
-
     width: 100%;
     max-width: 100%;
-    height: calc(100vh - ${HEADER_HEIGHT}px);
     max-height: 100%;
   `,
 }));
